### 2020-05-26 #2254 @whedon
[PRE REVIEW]: MorphoPy: A python package for feature extraction of neural morphologies.
**Submitting author:** @visdoom (<a href="http://orcid.org/0000-0001-9532-788X">Sophie Laturnus</a>)
**Repository:** <a href="https://github.com/berenslab/MorphoPy" target ="_blank">https://github.com/berenslab/MorphoPy</a>
**Version:** v0.6
