### 2020-05-26 f3ff7d7
Paper fixes
