### 2020-05-26 #2254 @danielskatz
👋 @Visdoom - you will need to remove the blank lines from the header of paper.md.  You also should change date to a date, following the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)

Once you've made these changes, please add a new comment here starting with ```@whedon generate pdf``` to see if the compilation works.  Iterate as needed.

(Also, Reference at the end should be References)
