### 2020-05-26 #2254 @arfon
@danielskatz @Visdoom - this PR should fix that build: https://github.com/berenslab/MorphoPy/pull/103 (it's got whitespace-alignment issues too)
