### 2020-05-26 6c1adca
Update paper.md
remove blank spaces in paper.md and update date
