+++ paper/morphopy.bib
@@ -58,12 +58,12 @@
   publisher={Cold Spring Harbor Laboratory}
 }
 
-@misc{NeuroM,
-  author = {​BBP},
+@misc{neurom,
+  author = {BBP},
   title = {NeuroM},
-  publisher = {​GitHub},
-  journal = {​GitHub repository},
-  url = {​https://github.com/BlueBrain/NeuroM}
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/BlueBrain/NeuroM}
 }
 
 @article{scorcioni:2008,
