### 2020-05-26 6c9d53a
Update paper.md
update NeuroM reference
