### 2020-05-26 b1d6d3b
Update paper.md
Fix citations
