### 2020-05-26 0569d09
Update paper.md
Moving figures in front of their reference in the text.
