### 2020-05-26 29f585f
Update paper.md
Fixed figure import?
