### 2020-05-26 7105139
Update paper.md
what is wrong with the figures?
