+++ paper/paper.md
@@ -74,12 +74,12 @@ Currently, MorphoPy supports the following feature representations:
 
 _Density maps_ are computed on the basis of a configuration file (or a dictionary) that controls parameters such as bin size
 and binning ranges. Additionally, users can specify whether and to which degree
-they want to smooth each density map \autoref{fig:dms}.
+they want to smooth each density map (see \autoref{fig:dms}).
 
 ![Node and edge related morphometric statistics. \label{fig:morphometrics}](./figures/fig_morphometrics.png)
 A variety of _morphometric statistics_ can be computed on the nodes and edges of each reconstruction.
 The `get_morphometric_statistics()`-method offers a precompiled single valued selection of these statistics including e.g.
-min/max branch angles, maximal branch order, and maximal path length to the soma (\autoref{fig:morphometrics}), but in principle, they can be adjusted to the user's personal preference.
+min/max branch angles, maximal branch order, and maximal path length to the soma (see \autoref{fig:morphometrics}), but in principle, they can be adjusted to the user's personal preference.
 
 Additionally, it is possible to query the entire _morphomoetric distribution_ of each statistic either in form of a histogram or as a
 Gaussian kernel density estimate (kde). Fig. \autoref{fig:morphdist}, for example, shows the kde of radial distances, branch angles and their
