### 2020-05-26 d9a0112
Update paper.md
still trying to fix image imports
