### 2020-05-26 3c10e63
Update paper.md
fixed typo
