### 2020-05-26 #2254 @Visdoom
> 👋 @Visdoom - Thanks for your submission to JOSS. As described in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss) announcing the reopening of JOSS, we're currently working in a "reduced service mode", limiting the number of papers assigned to any individual editor.
>
> Since reopening JOSS last week, we've had > 60 papers submitted and as such, yours has been put in our backlog that we will be working through over the coming weeks and months.
>
> Thanks in advance for your patience!
>
> (In the meantime, feel free to continue making fixes to the paper, adding missing DOIs, etc.)

Great, thank you!
