+++ paper/morphopy.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={91--96},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s12021-010-9093-7}
 }
 
 @article{li:2017,
@@ -16,7 +17,8 @@
   volume={12},
   number={8},
   year={2017},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0182184}
 }
 
 @article{kanari:2018,
@@ -27,7 +29,8 @@
   number={1},
   pages={3--13},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s12021-017-9341-1}
 }
 
 @article{laturnus:2019,
@@ -47,7 +50,8 @@
   number={1},
   pages={1--12},
   year={2019},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41467-019-12058-z}
 }
 
 @article{scala:2020,
@@ -55,7 +59,8 @@
   author={Scala, Federico and Kobak, Dmitry and Bernabucci, Matteo and Bernaerts, Yves and Cadwell, Cathryn R and Castro, Jesus R and Hartmanis, Leonard and Jiang, Xiaolong and Laturnus, Sophie R and Miranda, Elanine and others},
   journal={bioRxiv},
   year={2020},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/2020.02.03.929158}
 }
 
 @misc{neurom,
@@ -74,7 +79,8 @@
   number={5},
   pages={866},
   year={2008},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nprot.2008.51}
 }
 
 @article{jefferis:2007,
@@ -85,7 +91,8 @@
   number={6},
   pages={1187--1203},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cell.2007.01.040}
 }
 @techreport{hagberg:2008,
   title={Exploring network structure, dynamics, and function using NetworkX},
