### 2020-05-27 cb2c8ac
Moved figures after the text referencing them
