### 2020-05-27 094b74e
Use smaller image for 2D distribution
