### 2020-06-15 #2339 @oliviaguest
👋 👋 👋

Hi everybody!

@Visdoom, @emptymalei, @mstimberg — welcome to where the review itself is going to be.

Any questions please ask me. Any *very* code-related issues use the repo of the code and open an issue and link to it from here. All other issues, feedback, questions for me or the author, please use this issue. Anything you need, check above first as there is a lot of into up there, but also please feel free to ask me. 😸

Thank you all again and I hope this is a nice experience for us all! 🌼 🌸 🌷
