### 2020-06-21 berenslab/MorphoPy#104 @emptymalei
First example in the JOSS paper and the structure of the package
Thanks for this nice package. It is definitely very useful to anyone whoever has done the painful manual labor on determining the statistics. 😄

I am running into this problem regarding ~the installation process~ the first example in the paper. Would you mind looking into it? 😃

## What was tested

My platform:

1. OS: macOS High Sierra
2. Python: conda 4.8.0, python 3.5

I did the following:

1. Create a python 3.5 env: `conda create -n morphopy3.5 python=3.5 pip`
2. Install morphopy 0.6 in this env: `pip install morphopy`
3. `import MorphoPy.NeuronTree as nt` fails. I am following the demo [here](https://github.com/openjournals/joss-papers/blob/joss.02339/joss.02339/10.21105.joss.02339.pdf).

Here is the [link to the screencast](https://asciinema.org/a/r2dXMvNy8Q1WiogTclnBBc37f) showing the results.

[![asciicast](https://asciinema.org/a/r2dXMvNy8Q1WiogTclnBBc37f.svg)](https://asciinema.org/a/r2dXMvNy8Q1WiogTclnBBc37f)

In the screencast, we will see that we have no modules named `morphopy` or `MorphoPy`. Instead, we have `computation` and `neurontree`.

On the other hand, a small tweak of the example code [in the paper](https://github.com/openjournals/joss-papers/blob/joss.02339/joss.02339/10.21105.joss.02339.pdf) can make it work just fine.

```
import matplotlib.pyplot as plt
from computation import file_manager as fm
import neurontree as nt
from neurontree.plotting import show_threeview

N = fm.load_swc_file("~/Downloads/204-2-6nj.swc")
Dendrites = N.get_dendritic_tree()

fig = plt.figure(figsize=(10,10))
show_threeview(N, fig)
```

Maybe `import MorphoPy.NeuronTree as nt` is some legacy code?

## Why are `computation` and `neurontree` treated as top-level modules

The folder structure of this package is a bit different from the commonly used ones. As the folders `computation` and `neurontree` are top-level folders, they will be at the same level as other packages after installation, as shown in the above screencast.

I am not sure ~if this is by design~ why the package is designed in this way or maybe I missed the point.

If I may make any suggestions, I am used to a folder structure that has a folder `morphopy` with modules `computation` and `neurontree` in it. 😄

```
.
├── Creating\ density\ maps.ipynb
├── LICENSE
├── README.md
├── ...
├── morphopy
│    ├── MorphoPy.py
│    ├── __init__.py
│    ├── computation
│    └── neurontree
├── ...
└── ...
```

In this way, we will be able to `import morphopy.neurontree as nt` so that we won't interfere with modules in other packages since `computation` is kind of a generic module name.

Thank you very much for your time! Have a nice day! 🌞
