### 2020-06-21 berenslab/MorphoPy#105 @emptymalei
Typo: acylic
I believe the word `acyclic` was misspelt twice in the paper. 😸

I added comments to the markdown file so it is easier to locate the typos.

https://github.com/berenslab/MorphoPy/commit/83f147f7bc87d85feeadd955588a9f5748231997#diff-e2778e7674f45b806282a9611faa7220R52

https://github.com/berenslab/MorphoPy/commit/83f147f7bc87d85feeadd955588a9f5748231997#diff-e2778e7674f45b806282a9611faa7220R52
