### 2020-06-22 #2339 @oliviaguest
Oh, @emptymalei can you link to issues you open in the code repo here, please? Thank you! 😊

So far we have:

* https://github.com/berenslab/MorphoPy/issues/105
* https://github.com/berenslab/MorphoPy/issues/104

👍
