### 2020-06-22 berenslab/MorphoPy#104 @oliviaguest
Just a reminder, ignore if not applicable, that the bulk of the review (unless super technical) should be at: https://github.com/openjournals/joss-reviews/issues/2339 — thanks. ☺️
