### 2020-06-22 #2339 @emptymalei
@oliviaguest
I have tested most of the functions. Will finish this by the end of the week, too.
