### 2020-06-23 berenslab/MorphoPy#104 @Visdoom
Hi @emptymalei ,

thank you for being so thorough with the review.
This package structure is indeed a bit of legacy code and was kept because we did not want the file paths for the installation of the command-line tool. However, we agree that it is confusing and will change it as you have proposed.

@avdaranyi as we discussed, I will create the necessary folder structure and update all notebooks and code snippets. Once this is done, you can check and update the installation of the tool.
