### 2020-06-23 1fd1a58
Fixed Umlaut parsing in Tübingen
