### 2020-06-24 c2d8033
BUG FIX: to_swc. One does not need the data to be queried for the ids
