### 2020-06-25 #2339 @oliviaguest
@emptymalei thank you! Have you opened an issue in the code repo for this?

>  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
