### 2020-06-25 #2339 @emptymalei
@oliviaguest

I have a question here.
In fact, the code repo has a nicely illustrated tutorial jupyter notebook. It is neither unit test nor automated but it can serve as a very good manual test notebook.
I didn't check the box only because this is not automated. But based on this following section of the guidelines

> OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)

It is not bad.

In this case, shall I create an issue or only provide this as an feedback here?
