### 2020-06-25 berenslab/MorphoPy#106 @mstimberg
Some suggestions for the README file
(Issue opened as part of openjournals/joss-reviews#2339)

I found a few minor issues in the `README.md` file:
* Usage: "Just call everywhere..." sounds a bit odd, maybe "From anywhere, you can simply call.." or something similar?
* Citing MorphoPy: this should ideally more specific (telling the user exactly how to cite), but then I think this is what the JOSS paper is for :smile:
* Density maps: the function is misspelled as `compute_denisty_maps()` (same typo in the tutorial)
* Module description: it states that "If several soma points are present in the reconstruction file they will be automatically collapsed to the centroid of their convex hull on file loading", but I think this is only the case for > 3 points since MorphoPy also supports the 3-point soma format?
* Persistence: the text mentions how to implement custom persistence distance functions, but does not mention that if you use the API, you can directly provide this function as an argument. Inserting the function by editing a source file in the tool itself is not a great solution (most users won't even know where this file is after installing it with `pip`), but I completely understand that it would be quite an effort to make this work for the command line tool in another way.
