### 2020-06-25 berenslab/MorphoPy#107 @mstimberg
Some suggestions for the command line interface
(Issue opened as part of openjournals/joss-reviews#2339)

Testing the command line utility, I noticed that it handles a missing `-f` in a surprising way. If I want to run it on a file in a directory full of .swc files, but I accidentally forget the `-f` option, it will run on all files in the directory instead of giving an error message:
```console
$ MorphoPy.py -c stats ./C4.swc

['/mnt/data/anaconda2/envs/morphopy/bin/MorphoPy.py', '-c', 'stats', './C4.swc']
##### Morphometric statistics #####
... Process File:  EC3-80604.CNG.rotated.swc
   branch_points    width  ...  tree_asymmetry                   filename
0          216.0  1001.32  ...        0.814848  EC3-80604.CNG.rotated.swc

[1 rows x 29 columns]

... Process File:  EC3-80604.CNG.swc
There are 2-3 soma points. The location and the radius of the soma is estimated based on their mean.
[...]
```
Maybe one could even get rid of the `-f` and `-d` options completely, and have MorphoPy check whether the given name is a file or directory?

Additional minor suggestions for the command line utility:
* The tool always prints out `sys.argv` in the beginning, which looks a bit odd and is probably left-over from debugging?
* If you use incorrect options (or no options at all), you get a helpful error message but this message is quite hard to spot (see below). Maybe only print a short usage note for errors and have a `--help` /`-h` option for the full help?

![Screenshot](https://user-images.githubusercontent.com/1381982/85762889-7a75f080-b714-11ea-980b-ec1c6cf908da.png)
