### 2020-06-25 berenslab/MorphoPy#108 @mstimberg
Command line tool: incorrect handling of filename without path
(Issue opened as part of openjournals/joss-reviews#2339)

When running the command line tool under Linux, I noticed that it does not work correctly with filenames without a path.

E.g.:
```console
$ MorphoPy.py -c stats -f data/C4.swc  # all good
```
but
```console
$ cd data
$ MorphoPy.py -c stats -f C4.swc
[...]
FileNotFoundError: [Errno 2] File /C4.swc does not exist: '/C4.swc'
```

It seems to look for it at root instead of in the current directory (it works fine if I use `./C4.swc`).
