### 2020-06-25 #2339 @mstimberg
Ok, here's my review, adding to @emptymalei's remarks

I opened a few specific issues in the MorphoPy repository (berenslab/MorphoPy#106, berenslab/MorphoPy#107, berenslab/MorphoPy#108, berenslab/MorphoPy#109)

I agree with @emptymalei that the package structure is surprising and support changing it to a more standard organization.

I also wonder whether the API documentation could be made more accessible somehow? One option would be sphinx+readthedocs, but I see that the authors already have generated HTML documentation in `html_docs`. Maybe this could be put online in some way (Github pages, for example)?

 Regarding the test discussion:
 * Given that this tool computes statistics and can be used to gather these statistics automatically for large amounts of data, I think some kind of automated test suite would be really important. I'm sure that the authors checked the correctness of the statistical measurements extensively, but codifying this into some test would increase trust of potentially users. This would not have to be very complicated, I could imagine having an artificially created SWC file where the correct answers to all measurements are known and can therefore be directly compared to the results from the tool. Orthogonal to that would be some kind of regression test (to make sure that future changes to the software do not change the results). One possible "brute-force" solution would be to run the tool on all the files in the `data` directory and store the CSV result files in the repository as well; one could then easily run a script to check whether the results are still the same.
 * Testing that the plots are correct and do not change with changes in the software would of course be more complicated, and I do not think automatizing this in some way would be worth the effort. Having a tutorial notebook is a great solution for this part and also acts as useful documentation.
 * The same goes for the comparison to the TMD tool – I'm happy with a manual comparison in a notebook (but see berenslab/MorphoPy#109).

The software paper itself looks very good to me (apart from the Umlaut issue in the author affiliations, but that seems already having been corrected) – I wonder whether the code examples are a bit too technical, though? Maybe move them to the documentation (if they are not already mentioned there)?

All in all, great work, I think I'm going to use MorphoPy in the future myself!
