### 2020-06-26 berenslab/MorphoPy#108 @Visdoom
Hi @mstimberg,

thanks for pointing that out! I will try to fix that issue.
