### 2020-06-26 berenslab/MorphoPy#109 @Visdoom
Hi @mstimberg ,

good catch,  this error is caused by the changed API of the networkx package between 1.11 and 2.X.
I will update the functions to make them work with either version.
