+++ morphopy/neurontree/plotting.py
@@ -1,5 +1,3 @@
-
-from neurontree import NeuronTree
 import matplotlib.pyplot as plt
 from matplotlib_scalebar.scalebar import ScaleBar
 
