### 2020-06-26 6a5792a
Updated notebooks to work with new folder structure #104. Note! The sys imports have to be changed once the installation routine works properly again
