### 2020-06-26 041e952
Comparison notebook for TMD and MorphoPy can now handle both networkx versions. fixes #109
