+++ morphopy/MorphoPy.py
@@ -9,11 +9,10 @@ import scipy.io as sio
 import matplotlib as mpl
 mpl.use('Agg')
 import matplotlib.pyplot as plt
-from morphopy import computation as file_manager
+from morphopy import computation as file_manager, about
 import morphopy.computation.feature_presentation as fp
 import morphopy.computation.persistence_functions as pf
 # load global about variables
-import about
 # set global version
 __version__ = about.__version__
 
