### 2020-06-26 berenslab/MorphoPy#104 @Visdoom
@avdaranyi The folder structure is updated. Now you can go in and update the installation.
You can find the code in the branch https://github.com/berenslab/MorphoPy/tree/change_folder_structure
