### 2020-06-26 82561ef
Update code examples
The code examples in the paper now feature the updated package structure. #104
