### 2020-06-26 4f71545
distance functions independent of nx version. Closes #109
