### 2020-06-26 7b3bc94
Addressed the issues pointed out in #106. Fixes #106
