### 2020-06-26 berenslab/MorphoPy#106 @Visdoom
@mstimberg
thanks for all the thorough read of our README. In the following we list what we have changed.
1.  "Just call everywhere..." has been changed to "After the installation you can simply call"
2. The Citation paragraph now features a specific bibtex item. We will update it to the JOSS reference once this paper has been published.
3. typo is fixed
4. MorphoPy does not support the 3-point soma format (yet). When 2-3 points are present then their mean location is computed and taken as the soma. We updated the text accordingly.
5. We agree that writing into source files is not optimal but we decided to wait for the feedback of the community to see if that option is really used. We now mention the API option specifically.

With these updates we hope that it is fair to close this issue.
