### 2020-06-26 berenslab/MorphoPy#106 @mstimberg
> With these updates we hope that it is fair to close this issue.

I agree, all changes look good from my side.
