### 2020-06-26 berenslab/MorphoPy#106 @mstimberg
Oh, except that I just realized that it still uses `compute_denisty_maps()` instead of `compute_density_maps()`
