### 2020-06-26 #2339 @Visdoom
Hi @oliviaguest , @emptymalei  and @mstimberg ,

thank you for your reviewing efforts so far. Your feedback has been very helpful and your intend on future use @mstimberg is a great encouragement, thanks!
The remaining issues you have created via github will be addressed in the coming week. The biggest "issue" is to update the PyPi installation to the new folder structure but we are on it.

Regarding the automated test, we hope to provide both, the artificial test_swc morphometrics as well as the "brute-force" solution for checking the validity of future versions.
How is this handled though? Will I just let you know here when we have implemented that?

Best wishes, Sophie
