### 2020-06-26 berenslab/MorphoPy#106 @Visdoom
Are you sure? When I check in the repo I cannot see the typo:
https://github.com/berenslab/MorphoPy/blob/bd8a198d28e89197958d380e34ebf38f91d156e7/morphopy/computation/feature_presentation.py#L161
