### 2020-06-26 #2339 @oliviaguest
@Visdoom Hey Sophie, yes, you got off make any and all changes you fancy and report back here. Ideally if you can link from here or however you want highlight the changes you make, so the reviewers can easily spot them. Does that make sense? ☺️
