### 2020-06-26 berenslab/MorphoPy#106 @mstimberg
I meant in the README (sorry, there does not seem to be a better way to link to specific lines in a markdown file):
https://github.com/berenslab/MorphoPy/blame/master/README.md#L214

The same typo is also in the tutorial (section "Computing Density Maps")
