### 2020-06-28 d88ef57
Fixed bug: incorrect handling of filename
#108 filename is working also without leading slash now

Also updated setup.py: requirements only saved once in txt file
