+++ setup.py
@@ -5,8 +5,12 @@ import setuptools
 from morphopy import about
 
 # get long description from readme
-with open("README.md", "r", encoding='utf8') as fh:
-    long_description = fh.read()
+with open("README.md", "r", encoding='utf8') as f:
+    long_description = f.read()
+
+# get all requirements from txt file
+with open('requirements.txt') as f:
+    requirements = f.read().splitlines()
 
 setuptools.setup(
       name=about.__title__,
@@ -25,15 +29,5 @@ setuptools.setup(
          "License :: OSI Approved :: MIT License",
          "Operating System :: OS Independent",
       ],
-      install_requires=[
-         'matplotlib',
-         'matplotlib_scalebar',
-         'scipy',
-         'numpy',
-         'pandas',
-         'networkx',
-         'seaborn',
-         'sklearn',
-         'shapely'
-      ],
+      install_requires=requirements,
 )
