### 2020-06-28 berenslab/MorphoPy#108 @avdaranyi
Fixed with latest commit
