### 2020-06-28 berenslab/MorphoPy#104 @avdaranyi
The new folder structure is also adapted in the command line tool. "MorphoPy.py" needs to be outside of the package, because it is not part of it and only an external script file, which calls the package. So I moved it back outside. The next release will contain all changes.
