### 2020-06-28 berenslab/MorphoPy#104 @emptymalei
@avdaranyi

I have some ideas, if I may. 😄

Would it make sense to use the `console_scripts` approach instead of `scripts` approach?

Folder structure:

```
.
├── Creating\ density\ maps.ipynb
├── LICENSE
├── README.md
├── ...
├── morphopy
│    ├── MorphoPy.py
│    ├── __init__.py
│    ├── computation
│    └── neurontree
├── ...
└── ...
```

In `setup.py`

```
setup(
    ...
    entry_points = {
        'console_scripts': ['morphopy=morphopy.MorphoPy:main'],
    }
    ...
)
```

Then we can use the command `morphopy` in the terminal.

I never used the `scripts` approach but I know the `console_scripts` approach works when you need to import the package itself. Besides, the `console_scripts` approach also allows us to use the command line script in the actual python code.
