### 2020-06-28 #2339 @emptymalei
> I think I'm going to use MorphoPy in the future myself!

Also talked to my friend who's been doing such statistics by hand. She said this covers most of the statistics she's been looking at and will try out on the next project.
