### 2020-06-29 berenslab/MorphoPy#104 @avdaranyi
Hi @emptymalei

thank you for the suggestion. I think that's a very good idea. And on Windows an "exe"-file will be created which makes it easier to install there. With this approach I can run the command line script in python like this:

```
import morphopy.MorphoPy
morphopy.MorphoPy.main(['arg1', 'arg2', ...])
```

Thats also nice but do you know how can I run it on the command line without installing it?

If I'm inside the GitHub repository and want to test the command line tool outside of python then all imports will fail. So if I change it like this, I always have to switch to python command line and need to import the package first. That would make the testing a little bit ugly, also because I don't have the batch history and completion there.
