### 2020-06-29 berenslab/MorphoPy#107 @Visdoom
Hey @mstimberg,
@avdaranyi and I discussed the points you made above and this is what we came up with:
The first issue you rose deals with the `-f/-d` flag in our command-line tool. The default option is set to `-d` so this is why you found the directory to be read in even though no flag is provided. However, we agree that having to choose between the two input options `-f` for file or `-d` for a directory is unintuitive and prone to errors for the user. We, therefore, decided to merge this option into a `-i` for input flag. This flag will be mandatory so it will create an error if it is left out and the script will automatically detect the nature of the input (folder vs file).
3. We will slim down the error message to show the error and then a hint at how to open the --help interface. This should make the output easier to read.
