### 2020-06-29 65506ca
fixes #107
- removed -f and -d option from command line tool and replaced by -i/--input option which automatically recognizes file or directory input
- also slim help output on errors and long output only with --help
