+++ MorphoPy.py
@@ -9,8 +9,8 @@ import scipy.io as sio
 import matplotlib as mpl
 mpl.use('Agg')
 import matplotlib.pyplot as plt
-import morphopy.about as about
-import morphopy.computation.file_manager as file_manager
+from morphopy import about
+from morphopy.computation import file_manager
 import morphopy.computation.feature_presentation as fp
 import morphopy.computation.persistence_functions as pf
 # load global about variables
@@ -22,35 +22,39 @@ def help(exitcode=0):
     Print help page and exit application depending on error passed
     :param exitcode: Errorcode which will be returned at Exit or 0 if no error occured
     """
-    version()
+    if exitcode == 0:
+        version()
     print('')
-    print('Usage: MorphoPy.py -c <compute_feature> [--wide | --func <persistence_function> | --conf <config_file>]')
-    print('                   [-f <swc_file> | -d <directory>] [-o <output directory>]')
-    print('')
-    print('Options:')
-    print('   -c, --compute                parameter for selecting the computing feature:')
-    print('                                persistence: compute persistence data         ')
-    print('                                stats      : compute morphometric statistics  ')
-    print('                                density    : create density maps              ')
-    print('       statistics options:                                                    ')
-    print('       --long                   you can change your output format, in long    ')
-    print('                                format you get all values in a separate row.  ')
-    print('                                (default: all values in one row)              ')
-    print('       persistence options:                                                   ')
-    print('       --func                   if persistence is selected as feature, you can')
-    print('                                specify with this option a method function you')
-    print('                                want to use at computing the persistence.     ')
-    print('                                (default: radial distance function)           ')
-    print('       density map options:                                                   ')
-    print('       --conf                   if density map is selected, you can pass a    ')
-    print('                                config file with more parameters for creating ')
-    print('                                the density maps. (optional)                  ')
-    print('   -f, --file                   specifies a swc-file as input for Morphopy,   ')
-    print('                                if no file is selected, directory is used     ')
-    print('   -d, --directory              specifies a directory as input for swc-files. ')
-    print('                                (default: working directory)                  ')
-    print('   -o, --output                 specifies the output directory for saving the ')
-    print('                                results in. (default: same as source)         ')
+    print('Usage: MorphoPy.py -c <compute_feature> -i <swc_file>|<directory> [-o <output directory>]')
+    print('                   [--long | --func <persistence_function> | --conf <config_file>] [--help]')
+    print('                                                                                           ')
+    if exitcode == 0:
+        print('Options:                                                                      ')
+        print('   -h, --help                   show this help.                               ')
+        print('                                                                              ')
+        print('   -c, --compute                parameter for selecting the computing feature:')
+        print('                                persistence: compute persistence data         ')
+        print('                                stats      : compute morphometric statistics  ')
+        print('                                density    : create density maps              ')
+        print('       statistics options:                                                    ')
+        print('       --long                   you can change your output format, in long    ')
+        print('                                format you get all values in a separate row.  ')
+        print('                                (default: all values in one row)              ')
+        print('       persistence options:                                                   ')
+        print('       --func                   if persistence is selected as feature, you can')
+        print('                                specify with this option a method function you')
+        print('                                want to use at computing the persistence.     ')
+        print('                                (default: radial distance function)           ')
[…]
