### 2020-06-29 berenslab/MorphoPy#107 @mstimberg
All changes look good to me, I think the `-i` flag is a good solution.

> Secondly, the `sys.argv` is meant as a help for the user to see what parameters have been passed to the tool. As the command gets long it might lead to typos that are hard to spot. This is the reason why we'd prefer to keep it in the output but ultimately we do not have a strong opinion on this.

I don't have a strong opinion on this either, happy with leaving it in if that's what you prefer.
