### 2020-06-29 berenslab/MorphoPy#104 @emptymalei
@avdaranyi That's a good question. I don't really know how to do such tests except installing the package every time. :stuck_out_tongue:

On the other hand, this is not the most standard test in python. One possible solution is to write some `nose` tests and test the command line script. Then one only needs to do `nosetests` to test the script. I personally find the most useful tools to build a command-line tool is [`click`](https://click.palletsprojects.com/en/7.x/) which ships [runners to test the command-line tool itself](https://click.palletsprojects.com/en/7.x/testing/?highlight=test). This is only some personal experience.
