### 2020-07-01 63ddb9f
fixes #104 morphopy changed to console_scripts
MorphoPy.py is getting now a separate script during setup which calls morphopy as import. On Windows an exe-file will be generated. Call is much easier on command line. Args are passed to main function.

Documentation is also updated and everything ready for new release
