+++ setup.py
@@ -13,21 +13,23 @@ with open('requirements.txt') as f:
     requirements = f.read().splitlines()
 
 setuptools.setup(
-      name=about.__title__,
-      version=about.__version__,
-      scripts=['MorphoPy.py'],
-      author=about.__author__,
-      author_email=about.__email__,
-      description=about.__summary__,
-      long_description=long_description,
-      long_description_content_type="text/markdown",
-      url=about.__url__,
-      download_url="https://github.com/berenslab/MorphoPy/archive/v%s.tar.gz" % about.__version__,
-      packages=setuptools.find_packages(),
-      classifiers=[
-         "Programming Language :: Python :: 3",
-         "License :: OSI Approved :: MIT License",
-         "Operating System :: OS Independent",
-      ],
-      install_requires=requirements,
+    name=about.__title__,
+    version=about.__version__,
+    entry_points={
+        'console_scripts': ['morphopy=morphopy.MorphoPy:main'],
+    },
+    author=about.__author__,
+    author_email=about.__email__,
+    description=about.__summary__,
+    long_description=long_description,
+    long_description_content_type="text/markdown",
+    url=about.__url__,
+    download_url="https://github.com/berenslab/MorphoPy/archive/v%s.tar.gz" % about.__version__,
+    packages=setuptools.find_packages(),
+    classifiers=[
+     "Programming Language :: Python :: 3",
+     "License :: OSI Approved :: MIT License",
+     "Operating System :: OS Independent",
+    ],
+    install_requires=requirements,
 )
