### 2020-07-01 81b04c5
online documentation
new sphinx documentation for read the docs
format from all docstrings customied for sphinx
