+++ morphopy/neurontree/NeuronTree.py
@@ -21,9 +21,12 @@ matplotlib.rcParams.update({'font.size': 14})
 
 
 class NeuronTree:
-
-    # resample nodes along the edges of _G in distance d, return array of 3D positions
     def resample_nodes(self, d=1):
+        """resample nodes along the edges of _G in distance d, return array of 3D positions
+
+        :param d: distance
+        :return: array of positions
+        """
         P = []
         for (u, v, edata) in self._G.edges(data=True):
             if self._nxversion == 2:
@@ -46,18 +49,16 @@ class NeuronTree:
         P += list(self.get_node_attributes('pos').values())
         return np.array(P)
 
-    # creates a networkX Tree out of a swc file.
-    # scaling denotes the conversion factor needed to convert the units given in swc file to microns
-    # soma_rad denotes the radius of the soma given in microns
     def __init__(self, swc=None, scaling=1., node_data=[], edge_data=[], graph=None):
         """
         Creates a NeuronTree object that contains a networkx.DiGraph with node attributes 'pos' [x,y,z],
         'type' [1:soma,2:axon,3: dendrite], 'radius' and edge attributes 'euclidean_dist' and 'path_length'.
         It can be created from an swc file (as pandas.DataFrame or as numpy,ndarray) from a list of nodes and
         edges or from a given networkx.DiGraph.
+
         :param swc: pandas.DataFrame or numpy.ndarray containing the data of an swc file.
-        :param scaling: Scaling of the coordinates in swc file to microns. Sometimes the coordinates are given in voxels
-        or other units.
+        :param scaling: Scaling of the coordinates in swc file to microns.
+         Sometimes the coordinates are given in voxels or other units.
         :param soma_rad: radius of the soma in microns.
         :param node_data: list of nodes with node id and dictionary of node attributes
         :param edge_data: list of edges with dicionary of edge attributes
@@ -128,9 +129,9 @@ class NeuronTree:
             self._make_tree()   # needed to access the functions predecessor and successor
 
     def _remove_redundant_nodes(self):
-        """
-        Remove redundant nodes from the NeuronTree. A node is considered redundant if the edge between two nodes has a
-        euclidean distance = 0, so the node has the same 3D position as its predecessor.
+        """ Remove redundant nodes from the NeuronTree.
+         A node is considered redundant if the edge between two nodes has a euclidean distance = 0, so the node has
+          the same 3D position as its predecessor.
         """
         # get the nodes whose edge between them has distance = 0.
         nodeindices = np.array(list(self.get_edge_attributes('euclidean_dist').values())) == 0
@@ -163,8 +164,8 @@ class NeuronTree:
             edgelist = list(np.array(list(self.get_edge_attributes('euclidean_dist').keys()))[nodeindices])
 
     def _get_branch_type(self, B):
-        """
-        get the type of a branch based on majority vote.
+        """ get the type of a branch based on majority vote.
+
         :param B: subgraph, a branch within the NeuronTree
         :return: int, type id (1: soma, 2: axon, 3: basal dendrite, 4: apical dendrite). Type of the branch 'B'.
         """
@@ -230,7 +231,6 @@ class NeuronTree:
         looses the original node and edge attributes. Manipulation done inplace.
         changed for use with networkx v2 (works also in old version: parameters with names)
         """
-
         G = self._G
         roots = self.get_root(return_all=True) # in case there are disconnected neurites
 
@@ -264,8 +264,7 @@ class NeuronTree:
         Returns the topological minor of the Neuron. In this representation all continuation points are pruned away and
         the neuron solely consists of tips and branch points.
         Changed for use with networkx v2 (works also in old version: edge -> adj)
-        :return:
-            NeuronTree: mst. The topological minor representation of the original neuron.
[…]
