+++ morphopy/neurontree/utils.py
@@ -16,16 +16,8 @@ from scipy.spatial import ConvexHull
 def unit_vector(vector):
     """ Returns the unit vector of the vector.
 
-    Parameter
-    --------
-    vector : numpy.array
-        d-dimensional vector
-
-    Returns
-    -------
-    u : numpy.array
-        d-dimensional unit vector of 'vector'
-
+    :param vector: numpy.array d-dimensional vector
+    :return: u numpy.array d-dimensional unit vector of 'vector'
     """
     if np.linalg.norm(vector) > 0:
         return vector / np.linalg.norm(vector)
@@ -42,17 +34,10 @@ def angle_between(v1, v2):
             0.0
             >>> angle_between((1, 0, 0), (-1, 0, 0))
             3.141592653589793
-    Parameters
-    ----------
-    v1 : numpy.array
-        d-dimensional vector
-    v2 : numpy.array
-        d-dimensional vector
 
-    Returns
-    -------
-    theta : float
-        angle btw v1 and v2 in radians
+    :param v1: numpy.array d-dimensional vector
+    :param v2: numpy.array d-dimensional vector
+    :return: theta - float angle btw v1 and v2 in radians
 
     """
     v1_u = unit_vector(v1)
@@ -63,17 +48,9 @@ def angle_between(v1, v2):
 def get_axis(v1, v2):
     """ Returns the axis between two vectors v1 and v2
 
-    Parameters
-    ---------
-    v1: numpy.array
-        d-dimensional
-    v2: numpy.array
-        d-dimensional
-
-    Returns
-    -------
-    axis: numpy.array
-        d-dimensional axis btw v1 and v2
+    :param v1: numpy.array d-dimensional
+    :param v2: numpy.array d-dimensional
+    :return: axis - numpy.array d-dimensional axis btw v1 and v2
     """
     return (np.cross(v1, v2)) / (np.linalg.norm(np.cross(v1, v2)))
 
@@ -81,11 +58,9 @@ def get_axis(v1, v2):
 def get_A(x):
     """
     Returns matrix A. Used for rotation matrix calculation
-    :param x: 3D numpy.array
-            rotation axis
 
-    :return:
-        A: 3x3 numpy.array
+    :param x: 3D numpy.array rotation axis
+    :return: A - 3x3 numpy.array
     """
 
     if np.isnan(x).any():
@@ -105,7 +80,6 @@ def get_rotation_matrix(a, b):
     :param a: numpy.array (2 or 3 dimensional)
     :param b: numpy.array (2 or 3 dimensional)
     :return: R (2x2 or 3x3) rotation matrix to rotate a onto b
-
     """
 
     n = np.max(a.shape)
@@ -124,8 +98,8 @@ def get_rotation_matrix(a, b):
 
 
 def isRotationMatrix(R):
-    """
-    Checks if R is a valid rotation matrix.
+    """ Checks if R is a valid rotation matrix.
+
     :param R: [3x3] rotation matrix
     :return: boolean. Returns True when R is a valid rotation matrix, otherwise False.
     """
@@ -137,8 +111,8 @@ def isRotationMatrix(R):
 
 
 def rotationMatrixToEulerAngles(R):
-    """
-    Calculates rotation matrix to euler angles
+    """ Calculates rotation matrix to euler angles
+
     :param R: rotation matrix [3x3]
     :return: vector of euler angles (rotations around x,y and z axis)
     """
@@ -161,8 +135,8 @@ def rotationMatrixToEulerAngles(R):
 
 
 def eulerAnglesToRotationMatrix(theta):
-    """
-    Calculates the rotation matrix from given euelr angles
[…]
