### 2020-07-01 7e6a92b
Fix typo in density_map. Fix #106
