+++ tests/test_utils.py
@@ -1,134 +1,71 @@
 import numpy as np
-import sys
-
-sys.path.append('..')
 
 #### TEST GET_ANGLE #####
-from morphopy._utils.summarize import get_angle
+from morphopy.neurontree.utils import angle_between
 
 
 def test_get_angle_with_orthogonal_vectors():
     v0 = np.array([0, 0, 1])
     v1 = np.array([0, 1, 0])
 
-    r, d = get_angle(v0, v1)
+    r = angle_between(v0, v1)
     assert(r == 90*np.pi/180), "returned angle should be pi/2"
-    assert (d == 90), "returned angle should be 90 degree"
 
 
 def test_get_angle_with_opposite_vectors():
     v0 = np.array([0, 0, 1])
     v1 = np.array([0, 0, -1])
 
-    r, d = get_angle(v0, v1)
+    r= angle_between(v0, v1)
     assert (r == np.pi), "returned angle should be pi"
-    assert (d == 180), "returned angle should be 180 degree"
 
 
 def test_get_angle_with_same_vector():
     v0 = np.array([0, 0, 1])
 
-    r, d = get_angle(v0, v0)
+    r = angle_between(v0, v0)
     assert (r == 0), "returned angle should be 0"
-    assert (d == 0), "returned angle should be 0"
+
 
 
 def test_get_angle_with_unnormalized_vector():
     v0 = np.array([0, 0, 1])
     v1 = np.array([0, 2, 0])
 
-    r, d = get_angle(v0, v1)
+    r = angle_between(v0, v1)
     assert (r == 90 * np.pi / 180), "returned angle should be pi/2"
-    assert (d == 90), "returned angle should be 90 degree"
+
 
 
 def test_get_angle_btw_zero_and_v1():
     v0 = np.array([0, 0, 0])
     v1 = np.array([0, 1, 0])
 
-    r, d = get_angle(v0, v1)
+    r = angle_between(v0, v1)
     assert (r == 0), "returned angle should be 0"
-    assert (d == 0), "returned angle should be 0"
-
-
+    
+    
 def test_get_angle_returns_float():
     v0 = np.array([0, 0, 1])
     v1 = np.array([0, 1, 1])
 
-    r, d = get_angle(v0, v1)
+    r  = angle_between(v0, v1)
     assert (isinstance(r, np.float)), "get_angle() should return float"
-    assert (isinstance(d, np.float)), "get_angle() should return float"
+
 
 
 ### TEST READING METHODS ####
 
 import networkx as nx
-from morphopy._utils.utils import read_swc
+from morphopy.computation.file_manager import load_swc_file
 
 
 def test_read_swc_returned_fileformat():
 
     import pandas as pd
-    filepath = 'data/Image001-005-01.CNG.swc'
-    G, swc = read_swc(filepath)
+    filepath = '../data/Image001-005-01.CNG.swc'
+    G = load_swc_file(filepath)
 
-    assert(isinstance(G, nx.DiGraph)), "read_swc() should return a graph as networkx.DiGraph"
-    assert(isinstance(swc, pd.DataFrame)), "read_swc() should return a swc as pandas.DataFrame"
-
-
-def test_read_swc_all_variables_are_in():
-
-    filepath = 'data/Image001-005-01.CNG.swc'
-    G, swc = read_swc(filepath)
-
-    assert 'n' in swc.keys(), "column 'n' should be in pandas.DataFrame"
-    assert 'x' in swc.keys(), "column 'x' should be in pandas.DataFrame"
-    assert 'y' in swc.keys(), "column 'y' should be in pandas.DataFrame"
[…]
