### 2020-07-03 75b5877
set up test scripts
