### 2020-07-03 3a1927c
Test the current vs old morphometric statistics for 4 files
