### 2020-07-03 a4d0c44
Moved the file loading test to another file
