### 2020-07-03 8d198bd
New test files for persistence and density maps
