+++ tests/test_density_maps.py
@@ -1,1 +1,6 @@
 from morphopy.computation.file_manager import load_swc_file
+from morphopy.computation.feature_presentation import compute_density_maps
+
+file_names = ['Image001-005_01_CNenhance_latest_LXS.swc', 'C4.swc',
+              'V1_Layer23_Chat-IRES-Cre-neo_Ai14-299537.04.02.01_614430666_m.swc', 'EC3-80604.CNG.rotated.swc']
+
