### 2020-07-03 39024bc
test old against current computation of density maps
