### 2020-07-03 3cd087e
Added persistence tests for all four distance functions
