+++ .github/workflows/test.yml
@@ -22,6 +22,11 @@ jobs:
     # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
     - uses: actions/checkout@v2
 
+    - name: Install dependecies
+      run: |
+        python -m pip install --upgrade pip
+        pip install pytest
+        pip install morphopy
     # Runs a single command using the runners shell
     - name: Run a pytest
       run: pytest ./tests/
