+++ .github/workflows/test.yml
@@ -24,10 +24,13 @@ jobs:
 
     - name: Install dependecies
       run: |
-        python -m pip install --upgrade pip
+        python3 -m pip install --upgrade pip
         pip install pytest
         pip install morphopy
-    # Runs a single command using the runners shell
+
+    - name: Print current working dir
+      run: echo $PWD
+      
     - name: Run a pytest
       run: pytest ./tests/
 
