### 2020-07-03 9085ccb
Update test.yml
Test yml line by line
