+++ .github/workflows/test.yml
@@ -25,9 +25,9 @@ jobs:
     - name: Install dependencies
       run: |
         python3 -m pip install --upgrade pip
-        pip3 install pytest
-        pip3 install morphopy
-    - name: Shwo current working directory
+        python3 -m pip install pytest
+        python3 -m pip install morphopy
+    - name: Show current working directory
       run: echo $PWD
     - name: Run pytest 
       run: pytest ./tests/
