+++ .github/workflows/test.yml
@@ -26,6 +26,7 @@ jobs:
       run: |
         python3 -m pip install --upgrade pip
         python3 -m pip install pytest
+        python3 -m pip install setuptools
         python3 -m pip install morphopy
     - name: Show current working directory
       run: echo $PWD
