### 2020-07-03 ac83851
clean up code
