### 2020-07-03 5fb3ebb
clean up code
