+++ tests/test_persistence.py
@@ -23,7 +23,7 @@ def compare_old_and_current_persistence_implementation(filename, old_data, dista
     :param distance_function: distance function used for persistence data.
     """
 
-    neuron = load_swc_file("../data/%s" % filename)
+    neuron = load_swc_file(dir_path+"/../data/%s" % filename)
     columns = ['birth', 'death', 'node_type']
     old_persistence = old_data[old_data['filename'] == filename].sort_values(columns)[columns].reset_index()
     del old_persistence['index']
