### 2020-07-06 f82d6f6
updated html documentation
removed old and not needed html docs
restructured and added more docs for read the docs
