+++ docs/index.rst
@@ -1,19 +1,22 @@
 Welcome to Morphopy's documentation!
 ====================================
 
+MorphoPy is a Python3 package that uses networkX to compute and show information
+about neurites. The input can be passed with single swc-files or it can handle
+whole directories with multiple files at once. You can use MorphoPy imported in
+Python or from command line as batch-tool as well.
+
+You can find MorphoPy with more information `here... <https://github.com/berenslab/MorphoPy>`_
+
+Contents:
+=========
+
 .. toctree::
    :maxdepth: 2
-   :caption: Contents:
 
-   morphopy.rst
-   
-   feature_presentation.rst
-   file_manager.rst
-   persistence_functions.rst
-   
-   neurontree.rst
-   plotting.rst
-   utils.rst
+   command.rst
+   computation.rst
+   neuron.rst
 
 Indices and tables
 ==================
