+++ docs/neuron.rst
@@ -1,0 +1,8 @@
+NeuronTree
+----------
+.. toctree::
+   :maxdepth: 2
+
+   neurontree.rst
+   plotting.rst
+   utils.rst
