### 2020-07-07 #2339 @Visdoom
Hey @oliviaguest,
thanks for checking in. We have created some tests but the automated integration into gitHub has a hick-up that I need to fix. I meant to do it yesterday and then ran out of time. So we are almost there. I will give you a wrap up here once everything has been addressed.
