### 2020-07-07 1d5bb94
BUG FIX: branch_order did not work with networkx v2
