+++ .github/workflows/test.yml
@@ -27,7 +27,7 @@ jobs:
         python3 -m pip install --upgrade pip
         python3 -m pip install pytest
         python3 -m pip install setuptools
-        python3 -m pip install morphopy
+        python3 -m pip install git+https://github.com/berenslab/MorphoPy
     - name: Run pytest 
       run: /home/runner/.local/bin/pytest ./tests/
 
