### 2020-07-07 5575c5e
reduce neurons before computing the persistence
