+++ .github/workflows/test.yml
@@ -26,7 +26,7 @@ jobs:
       run: |
         python3 -m pip install --upgrade pip
         python3 -m pip install pytest
-        python3 -m pip install setuptools
+        python3 -m pip install setuptools==41.0.1
         python3 -m pip install git+https://github.com/berenslab/MorphoPy
     - name: Run pytest 
       run: /home/runner/.local/bin/pytest ./tests/
