### 2020-07-07 b72668b
Update test.yml
Force python version to be 3.5
