+++ .github/workflows/test.yml
@@ -21,12 +21,17 @@ jobs:
     steps:
     # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
     - uses: actions/checkout@v2
+    
+    - name: Set up Python 3.5
+      uses: actions/setup-python@v1
+      with:
+        python-version: 3.5
 
     - name: Install dependencies
       run: |
         python3 -m pip install --upgrade pip
         python3 -m pip install pytest
-        python3 -m pip install setuptools==41.0.1
+        python3 -m pip install setuptools
         python3 -m pip install git+https://github.com/berenslab/MorphoPy
     - name: Run pytest 
       run: /home/runner/.local/bin/pytest ./tests/
