### 2020-07-07 de4d576
Update test.yml
Remove path to pytest
