### 2020-07-07 1520798
Update paper.md
Update command line call and add read the docs link
