+++ paper/paper.md
@@ -115,9 +115,9 @@ In addition to working as a package in interactive computing environments, `Morp
 to operate on single files or entire batches.
 ```bash
 MorphoPy.py -c [density|persistence|stats] 
-            -f ['path_to_file'] | -d ['path_to_folder']
+            -i ['path_to_file'|'path_to_folder']
 ```
-For a full documentation of `MorphoPy`'s functionality please refer to our documentation and tutorial on our [gitHub page](https://github.com/berenslab/MorphoPy).
+For a full documentation of `MorphoPy`'s functionality please refer to our [documentation](https://morphopy.readthedocs.io/en/latest/) and tutorial on our [gitHub page](https://github.com/berenslab/MorphoPy).
 
 `MorphoPy` has been developed in the context of a benchmarking study for cortical interneuron cell type classification
 based on their morphology [@laturnus:2019]. It has already been used in a series of scientific publications that relate
