### 2020-07-09 a9ff45c
BUG Fix: get_path_angle() depends on sorting of successor array. Fixes #112
