### 2020-07-09 4cd7672
Install newest Python version and adjust path to pytest
