+++ .github/workflows/test.yml
@@ -22,11 +22,6 @@ jobs:
     # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
     - uses: actions/checkout@v2
     
-    - name: Set up Python 3.5
-      uses: actions/setup-python@v1
-      with:
-        python-version: 3.5
-
     - name: Install dependencies
       run: |
         python3 -m pip install --upgrade pip
@@ -34,5 +29,5 @@ jobs:
         python3 -m pip install setuptools
         python3 -m pip install git+https://github.com/berenslab/MorphoPy
     - name: Run pytest 
-      run: pytest ./tests/
+      run: /home/runner/.local/bin/pytest ./tests/
 
