### 2020-07-10 #2339 @Visdoom
Dear @oliviaguest, @emptymalei and @mstimberg,
again, thank you for taking the time to review the code and the paper to our package MorphoPy. Your input has been valuable and we believe, that the package has improved majorly. The most prominent change has been the updated folder structure where now `morphopy` is the parent folder to all functional subfolders.
 |-- morphopy
    |-- computations
    |-- neurontree
    |-- about
….
The new importing structure that this implies has been updated for all notebooks and code-snippets within the paper.
Furthermore, the full documentation of the package is now available under  https://morphopy.readthedocs.io/en/latest/.
Additionally, we have added a testing suite in the the folder `tests`, which is run every time a push or a pull request is performed for the master branch on gitHub.
Within these tests we mainly check if the current implementation of feature representations is consistent with an older precomputed version that is stored in cvs-files in `tests/test_data/` for four realistic neuron files.
If you as reviewers want to run the tests separately you can do that by cloning the repository and running `pytest ./tests/` from the main folder - note, that `pytest` has to be installed for that.
Lastly, all the gitHub issues raised by the reviewers have been fixed.
Please be aware, that the updated package version is not available on PyPi yet. We will be releasing it soon as version 0.7.1 in case you want to recheck via the pip installation.

Thank you very much and we hope we have addressed all your concerns,
Team Berenslab
