### 2020-07-10 #2339 @mstimberg
You put some serious work into these improvements, that's great to see. I am very happy with the state of the package now, and I can confirm that the test suite runs fine on my machine. A few minor issues:
* The JOSS manuscript still describes the outdated `-f`/`-d` syntax for the command line tool
*  The documentation on readthedocs should probably make it a bit clearer that it is only the API documentation and that the more high-level general introduction is in the README on github (or include this documentation directly), also not sure whether you link to readthedocs anywhere?
* In the readthedocs documentation, [Command Line Script](https://morphopy.readthedocs.io/en/latest/command.html) links to an empty page.

From my side, I'll be happy to see this published as it is (pending the minor changes suggested above, in particular the first point). @oliviaguest: I will be out of office for the three coming weeks, but consider this my formal publication endorsement, no need to block the publication until you hear back from me.

Thanks and bye everyone :wave:

PS: A mere suggestion because I saw that some tests were quite redundant (testing the same thing for several files, etc.). You might want to have a look at pytest's [parametrize](https://docs.pytest.org/en/latest/parametrize.html#pytest-mark-parametrize-parametrizing-test-functions) decorator that makes it possible to express such tests more concisely.
