### 2020-07-10 #2339 @emptymalei
Great work. Checked all the items in the checklist. Agree to publish. 👍

Thank you for this useful package.

Two possible minor improvements for later updates maybe. I believe these should not be blockers to the paper itself.

The first one is for the automated test. I had a look at the Actions workflow.
[This line in the workflow](https://github.com/berenslab/MorphoPy/blob/4cd7672d4f27909e90f671a2c9d8b9d0348257cf/.github/workflows/test.yml#L30)
```
         python3 -m pip install git+https://github.com/berenslab/MorphoPy
```
would only test the remote master branch. In this case, it works but a bit strange since you already have exactly the same files locally.

The test actually could be done for [all branches on a matrix of platforms](https://docs.github.com/en/actions/language-and-framework-guides/using-python-with-github-actions#excluding-a-version). When the `checkout` action is used, you actually have all the content on the action server "locally" so you could install and test right using all the local file. Also, the test can also be done on a matrix of platforms and many different versions of python.

The second one is about the versions of requirements. I should have mentioned this earlier. It is considered a good habit to pin the package versions in the `requirements.txt` file. For example, networkx has many breaking changes in different versions (such as members of DiGraphView) and it would be more stable if one could pin the version or range of versions in `requirements.txt`. (From a data scientist's experience, `requirements.txt` without version restrictions will drive our data dev ops crazy. 😸  )
