### 2020-07-12 70445a7
Updates for new Morphopy Version
added Link in doc to readthedocs,
and new morphopy version after patch
