### 2020-07-21 #2339 @oliviaguest
@Visdoom ah, also I spotted a tiny typo: https://github.com/berenslab/MorphoPy/pull/113

When you merge that feel free to ask @whedon to recompile the PDF. 😸
