### 2020-07-24 #2339 @oliviaguest
@Visdoom is this PDF looking good to you, shall we move to acceptance? ☺️
