### 2020-08-03 #2339 @Visdoom
@oliviaguest I noticed a minor issue in the text where it says Fig. Figure X. Can I still update it? Otherwise it is also not a tragedy.
