### 2020-08-03 fd3b54a
Update paper.md
Remove superfluous 'Fig.'
