### 2020-08-03 #2339 @Visdoom
Great, it is fixed in the 'paper.md'.
