### 2020-08-03 cd97ed8
minor bib changes
