+++ paper/morphopy.bib
@@ -13,7 +13,7 @@
 @article{li:2017,
   title={Metrics for comparing neuronal tree shapes based on persistent homology},
   author={Li, Yanjie and Wang, Dingkang and Ascoli, Giorgio A and Mitra, Partha and Wang, Yusu},
-  journal={PloS one},
+  journal={PLOS ONE},
   volume={12},
   number={8},
   year={2017},
@@ -45,7 +45,7 @@
 @article{scala:2019,
   title={Layer 4 of mouse neocortex differs in cell types and circuit organization between sensory areas},
   author={Scala, Federico and Kobak, Dmitry and Shan, Shen and Bernaerts, Yves and Laturnus, Sophie and Cadwell, Cathryn Rene and Hartmanis, Leonard and Froudarakis, Emmanouil and Castro, Jesus Ramon and Tan, Zheng Huan and others},
-  journal={Nature communications},
+  journal={Nature Communications},
   volume={10},
   number={1},
   pages={1--12},
@@ -74,7 +74,7 @@
 @article{scorcioni:2008,
   title={L-Measure: a web-accessible tool for the analysis, comparison and search of digital reconstructions of neuronal morphologies},
   author={Scorcioni, Ruggero and Polavaram, Sridevi and Ascoli, Giorgio A},
-  journal={Nature protocols},
+  journal={Nature Protocols},
   volume={3},
   number={5},
   pages={866},
@@ -95,7 +95,7 @@
   doi={10.1016/j.cell.2007.01.040}
 }
 @techreport{hagberg:2008,
-  title={Exploring network structure, dynamics, and function using NetworkX},
+  title={Exploring network structure, dynamics, and function using {NetworkX}},
   author={Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
   year={2008},
   institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
