+++ paper/morphopy.bib
@@ -98,5 +98,5 @@
   title={Exploring network structure, dynamics, and function using {NetworkX}},
   author={Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
   year={2008},
-  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
+  institution={Los Alamos National Lab. (LANL), Los Alamos, NM, United States}
 }
