### 2020-08-03 #2339 @danielskatz
@Visdoom  - please merge https://github.com/berenslab/MorphoPy/pull/114 before we proceed to acceptance
