### 2020-08-03 #2339 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
